+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file based on the
 - Function score query: corrected the `score_method` `average` to `avg` #975 
 - Set `json_decode()` assoc parameter to true in `Elastica\Response` #1005
 - Add `bigintConversion` to keys passed to connection config in `Elastica\Client` #1005
+- Use POST instead of PUT to send bulk requests #1010
 
 ### Added
 - Elastica\Query\MultiMatch::setFuzziness now supports being set to `AUTO` with the const `MultiMatch::FUZZINESS_AUTO`